+++ app/scripts/fhir/STU3/lformsFHIR.js
@@ -22345,6 +22345,13 @@ var self = {
 
         break;
 
+      case "attachment":
+        values = [{
+          key: "valueAttachment",
+          val: item.value
+        }];
+        break;
+
       default:
         values = [{
           key: "valueString",
@@ -22954,6 +22961,12 @@ var self = {
 __webpack_require__.r(__webpack_exports__);
 function _defineProperty(obj, key, value) { if (key in obj) { Object.defineProperty(obj, key, { value: value, enumerable: true, configurable: true, writable: true }); } else { obj[key] = value; } return obj; }
 
+function _createForOfIteratorHelper(o, allowArrayLike) { var it; if (typeof Symbol === "undefined" || o[Symbol.iterator] == null) { if (Array.isArray(o) || (it = _unsupportedIterableToArray(o)) || allowArrayLike && o && typeof o.length === "number") { if (it) o = it; var i = 0; var F = function F() {}; return { s: F, n: function n() { if (i >= o.length) return { done: true }; return { done: false, value: o[i++] }; }, e: function e(_e) { throw _e; }, f: F }; } throw new TypeError("Invalid attempt to iterate non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method."); } var normalCompletion = true, didErr = false, err; return { s: function s() { it = o[Symbol.iterator](); }, n: function n() { var step = it.next(); normalCompletion = step.done; return step; }, e: function e(_e2) { didErr = true; err = _e2; }, f: function f() { try { if (!normalCompletion && it.return != null) it.return(); } finally { if (didErr) throw err; } } }; }
+
+function _unsupportedIterableToArray(o, minLen) { if (!o) return; if (typeof o === "string") return _arrayLikeToArray(o, minLen); var n = Object.prototype.toString.call(o).slice(8, -1); if (n === "Object" && o.constructor) n = o.constructor.name; if (n === "Map" || n === "Set") return Array.from(o); if (n === "Arguments" || /^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n)) return _arrayLikeToArray(o, minLen); }
+
+function _arrayLikeToArray(arr, len) { if (len == null || len > arr.length) len = arr.length; for (var i = 0, arr2 = new Array(len); i < len; i++) { arr2[i] = arr[i]; } return arr2; }
+
 /**
  *  Defines SDC export functions that are the same across the different FHIR
  *  versions.  The function takes the SDC namespace object defined in the sdc export
@@ -23187,6 +23200,35 @@ function addCommonSDCExportFns(ns) {
         targetItem.item.push(helpItem);
       } else {
         targetItem.item = [helpItem];
+      }
+    }
+
+    if (item.maxAttachmentSize) {
+      var exts = targetItem.extension || (targetItem.extension = []);
+      exts.push({
+        url: self.fhirExtMaxSize,
+        valueDecimal: item.maxAttachmentSize
+      });
+    }
+
+    if (item.allowedAttachmentTypes) {
+      exts = targetItem.extension || (targetItem.extension = []);
+
+      var _iterator = _createForOfIteratorHelper(item.allowedAttachmentTypes),
+          _step;
+
+      try {
+        for (_iterator.s(); !(_step = _iterator.n()).done;) {
+          var type = _step.value;
+          exts.push({
+            url: self.fhirExtMimeType,
+            valueCode: type
+          });
+        }
+      } catch (err) {
+        _iterator.e(err);
+      } finally {
+        _iterator.f();
       }
     } // handle special constraints for "display" item
 
@@ -23771,12 +23813,12 @@ function addCommonSDCExportFns(ns) {
    * Here are the details for a single value's conversion (to an element in the returned answer array)
    * - For item data type quantity (QTY), a valueQuantity answer element will be created IF
    *   either (or both) item value or item unit is available.
-   * - For item data types boolean, decimal, integer, date, dateTime, instant, time, string, and url,
+   * - For item data types boolean, decimal, integer, date, dateTime, instant, time, string, attachment, and url,
    *   it will be converted to a FHIR value{TYPE} entry if the value is not null, not undefined, and not
    *   an empty string.
    * - For CNE and CWE, a valueCoding entry is created IF at least one of the item value's code, text, or system
    *   is available
-   * - No answer entry will be created in all other cases, e.g., for types reference, title, section, attachment, etc.
+   * - No answer entry will be created in all other cases, e.g., for types reference, title, section, etc.
    * @param item the item whose value is to be converted
    * @return the converted FHIR QuestionnaireResponse answer (an array), or null if the value is not converted -
    *         see the function description above for more details.
@@ -23824,8 +23866,8 @@ function addCommonSDCExportFns(ns) {
             //   "code" : "<code>" // Coded form of the unit
             // }]
             answer = this._setIfHasValue(null, 'valueQuantity', this._makeValueQuantity(itemValue, item.unit));
-          } // for boolean, decimal, integer, date, dateTime, instant, time, string, uri
-          else if (dataType === "BL" || dataType === "REAL" || dataType === "INT" || dataType === "DT" || dataType === "DTM" || dataType === "TM" || dataType === "ST" || dataType === "TX" || dataType === "URL") {
+          } // for boolean, decimal, integer, date, dateTime, instant, time, string, uri, attachment
+          else if (this._lformsTypesToFHIRFields[dataType]) {
               var valueKey = this._getValueKeyByDataType("value", item);
 
               answer = _defineProperty({}, valueKey, itemValue);
@@ -24529,112 +24571,6 @@ function addSDCImportFns(ns) {
       }
 
       return item;
-    },
-
-    /**
-     * Set value and units on a LForms item
-     * @param linkId an item's linkId
-     * @param answer value for the item
-     * @param item a LForms item
-     * @private
-     */
-    _setupItemValueAndUnit: function _setupItemValueAndUnit(linkId, answer, item) {
-      if (item && linkId === item.linkId && item.dataType !== 'SECTION' && item.dataType !== 'TITLE') {
-        var dataType = item.dataType; // any one has a unit must be a numerical type, let use REAL for now.
-        // dataType conversion should be handled when panel data are added to lforms-service.
-
-        if ((!dataType || dataType === "ST") && item.units && item.units.length > 0) {
-          item.dataType = dataType = "REAL";
-        }
-
-        var qrValue = answer[0];
-
-        switch (dataType) {
-          case "BL":
-            if (qrValue.valueBoolean === true || qrValue.valueBoolean === false) {
-              item.value = qrValue.valueBoolean;
-            }
-
-            break;
-
-          case "INT":
-            if (qrValue.valueQuantity) {
-              item.value = qrValue.valueQuantity.value;
-
-              if (qrValue.valueQuantity.code) {
-                item.unit = {
-                  name: qrValue.valueQuantity.code
-                };
-              }
-            } else if (qrValue.valueInteger) {
-              item.value = qrValue.valueInteger;
-            }
-
-            break;
-
-          case "REAL":
-          case "QTY":
-            if (qrValue.valueQuantity) {
-              item.value = qrValue.valueQuantity.value;
-
-              if (qrValue.valueQuantity.code) {
-                item.unit = {
-                  name: qrValue.valueQuantity.code
-                };
-              }
-            } else if (qrValue.valueDecimal) {
-              item.value = qrValue.valueDecimal;
-            }
-
-            break;
-
-          case "DT":
-            item.value = qrValue.valueDate;
-            break;
-
-          case "DTM":
-            item.value = qrValue.valueDateTime;
-            break;
-
-          case "CNE":
-          case "CWE":
-            if (ns._answerRepeats(item)) {
-              var value = [];
-
-              for (var j = 0, jLen = answer.length; j < jLen; j++) {
-                var val = ns._processCWECNEValueInQR(answer[j]);
-
-                if (val) {
-                  value.push(val);
-                }
-              }
-
-              item.value = value;
-            } else {
-              var val = ns._processCWECNEValueInQR(qrValue);
-
-              if (val) {
-                item.value = val;
-              }
-            }
-
-            break;
-
-          case "ST":
-          case "TX":
-            item.value = qrValue.valueString;
-            break;
-
-          case "SECTION":
-          case "TITLE":
-          case "":
-            // do nothing
-            break;
-
-          default:
-            item.value = qrValue.valueString;
-        }
-      }
     }
   };
 }
@@ -24682,11 +24618,13 @@ function addCommonSDCFns(ns) {
     "URL": 'url',
     "CNE": 'choice',
     "CWE": 'open-choice',
-    "QTY": 'quantity'
+    "QTY": 'quantity',
+    "attachment": 'attachment'
   }; // A mapping from LHC-Forms data types to the partial field names of the value fields
   // and initial value fields in FHIR Questionnaire
 
   self._lformsTypesToFHIRFields = {
+    "attachment": "Attachment",
     "INT": 'Integer',
     "REAL": 'Decimal',
     "DT": 'Date',
@@ -25169,10 +25107,27 @@ function addCommonSDCImportFns(ns) {
   self.fhirExtAnswerExp = "http://hl7.org/fhir/uv/sdc/StructureDefinition/sdc-questionnaire-answerExpression";
   self.fhirExtChoiceOrientation = "http://hl7.org/fhir/StructureDefinition/questionnaire-choiceOrientation";
   self.fhirExtLaunchContext = "http://hl7.org/fhir/uv/sdc/StructureDefinition/sdc-questionnaire-launchContext";
+  self.fhirExtMaxSize = "http://hl7.org/fhir/StructureDefinition/maxSize";
+  self.fhirExtMimeType = "http://hl7.org/fhir/StructureDefinition/mimeType";
   self.fhirExtUrlRestrictionArray = [self.fhirExtUrlMinValue, self.fhirExtUrlMaxValue, self.fhirExtUrlMinLength, self.fhirExtUrlRegex]; // One way or the other, the following extensions are converted to lforms internal fields.
   // Any extensions not listed here (there are many) are recognized as lforms extensions as they are.
 
-  self.handledExtensionSet = new Set([self.fhirExtUrlCardinalityMin, self.fhirExtUrlCardinalityMax, self.fhirExtUrlItemControl, self.fhirExtUrlUnit, self.fhirExtUrlUnitOption, self.fhirExtUrlOptionPrefix, self.fhirExtUrlMinValue, self.fhirExtUrlMaxValue, self.fhirExtUrlMinLength, self.fhirExtUrlRegex, self.fhirExtUrlAnswerRepeats, self.fhirExtUrlExternallyDefined, self.argonautExtUrlExtensionScore, self.fhirExtUrlHidden, self.fhirExtTerminologyServer, self.fhirExtUrlDataControl, self.fhirExtChoiceOrientation]);
+  self.handledExtensionSet = new Set([self.fhirExtUrlCardinalityMin, self.fhirExtUrlCardinalityMax, self.fhirExtUrlItemControl, self.fhirExtUrlUnit, self.fhirExtUrlUnitOption, self.fhirExtUrlOptionPrefix, self.fhirExtUrlMinValue, self.fhirExtUrlMaxValue, self.fhirExtUrlMinLength, self.fhirExtUrlRegex, self.fhirExtUrlAnswerRepeats, self.fhirExtUrlExternallyDefined, self.argonautExtUrlExtensionScore, self.fhirExtUrlHidden, self.fhirExtTerminologyServer, self.fhirExtUrlDataControl, self.fhirExtChoiceOrientation]); // Simple functions for mapping extensions to properties in the internal structure.
+  // Parameters:
+  //   extension: the FHIR extension object
+  //   item:  The LForms item to be updated
+
+  self.extensionHandlers = {};
+
+  self.extensionHandlers[self.fhirExtMaxSize] = function (extension, item) {
+    item.maxAttachmentSize = extension.valueDecimal || extension.valueInteger; // not sure why it is decimal
+  };
+
+  self.extensionHandlers[self.fhirExtMimeType] = function (extension, item) {
+    item.allowedAttachmentTypes || (item.allowedAttachmentTypes = []);
+    item.allowedAttachmentTypes.push(extension.valueCode);
+  };
+
   self.formLevelFields = [// Resource
   'id', 'meta', 'implicitRules', 'language', // Domain Resource
   'text', 'contained', 'extension', 'modifiedExtension', // Questionnaire
@@ -25858,6 +25813,117 @@ function addCommonSDCImportFns(ns) {
     }
   };
   /**
+   * Set value and units on a LForms item
+   * @param linkId an item's linkId
+   * @param answer value for the item
+   * @param item a LForms item
+   * @private
+   */
+
+
+  qrImport._setupItemValueAndUnit = function (linkId, answer, item) {
+    if (item && linkId === item.linkId && item.dataType !== 'SECTION' && item.dataType !== 'TITLE') {
+      var dataType = item.dataType; // any one has a unit must be a numerical type, let use REAL for now.
+      // dataType conversion should be handled when panel data are added to lforms-service.
+
+      if ((!dataType || dataType === "ST") && item.units && item.units.length > 0) {
+        item.dataType = dataType = "REAL";
+      }
+
+      var qrValue = answer[0];
+
+      switch (dataType) {
+        case "BL":
+          if (qrValue.valueBoolean === true || qrValue.valueBoolean === false) {
+            item.value = qrValue.valueBoolean;
+          }
+
+          break;
+
+        case "INT":
+          if (qrValue.valueQuantity) {
+            item.value = qrValue.valueQuantity.value;
+
+            if (qrValue.valueQuantity.code) {
+              item.unit = {
+                name: qrValue.valueQuantity.code
+              };
+            }
+          } else if (qrValue.valueInteger) {
+            item.value = qrValue.valueInteger;
+          }
+
+          break;
+
+        case "REAL":
+        case "QTY":
+          if (qrValue.valueQuantity) {
+            item.value = qrValue.valueQuantity.value;
+
+            if (qrValue.valueQuantity.code) {
+              item.unit = {
+                name: qrValue.valueQuantity.code
+              };
+            }
+          } else if (qrValue.valueDecimal) {
+            item.value = qrValue.valueDecimal;
+          }
+
+          break;
+
+        case "DT":
+          item.value = qrValue.valueDate;
+          break;
+
+        case "DTM":
+          item.value = qrValue.valueDateTime;
+          break;
+
+        case "CNE":
+        case "CWE":
+          if (ns._answerRepeats(item)) {
+            var value = [];
+
+            for (var j = 0, jLen = answer.length; j < jLen; j++) {
+              var val = ns._processCWECNEValueInQR(answer[j]);
+
+              if (val) {
+                value.push(val);
+              }
+            }
+
+            item.value = value;
+          } else {
+            var val = ns._processCWECNEValueInQR(qrValue);
+
+            if (val) {
+              item.value = val;
+            }
+          }
+
+          break;
+
+        case "ST":
+        case "TX":
+          item.value = qrValue.valueString;
+          break;
+
+        case "attachment":
+          item.value = qrValue.valueAttachment;
+          break;
+
+        case "SECTION":
+        case "TITLE":
+        case "":
+          // do nothing
+          break;
+
+        default:
+          item.value = qrValue.valueString;
+      }
+    }
+  };
+  /**
    * Get LForms data type from questionnaire item
    *
    * @param qItem {object} - Questionnaire item object
@@ -25924,6 +25990,10 @@ function addCommonSDCImportFns(ns) {
 
       case "quantity":
         type = 'QTY';
+        break;
+
+      case "attachment":
+        type = 'attachment';
         break;
     }
 
@@ -26274,7 +26344,9 @@ function addCommonSDCImportFns(ns) {
 
     if (Array.isArray(qItem.extension)) {
       for (var i = 0; i < qItem.extension.length; i++) {
-        if (!self.handledExtensionSet.has(qItem.extension[i].url)) {
+        var ext = qItem.extension[i];
+        var extHandler = self.extensionHandlers[ext.url];
+        if (extHandler) extHandler(ext, lfItem);else if (!self.handledExtensionSet.has(qItem.extension[i].url)) {
           extensions.push(qItem.extension[i]);
         }
       }