+++ app/scripts/lib/lforms-data.js
@@ -2024,7 +2024,10 @@
         }
         // single selection answer list
         else if (angular.isObject(item.value)) {
-          isEmpty = item.value.text === undefined || item.value.text === null || item.value.text ==="";
+          // An attachment should have either either value.data or value.url,
+          var v = item.value;
+          isEmpty = (v.data == undefined || v.data == null) && !v.url && // attachment
+            !v.text; // !v.text means !(undefined or null or "")
         }
         // simple type
         else if (item.value !== undefined && item.value !== null && item.value !=="") {